+++ rtic/src/export.rs
@@ -12,41 +12,37 @@ pub use cortex_common::*;
 mod cortex_common;
 
 // Cortex-M target with basepri support
-#[cfg(any(feature = "cortex-m-basepri", feature = "rtic-uitestv7"))]
+#[cfg(feature = "cortex-m-basepri")]
 mod cortex_basepri;
 
-#[cfg(any(feature = "cortex-m-basepri", feature = "rtic-uitestv7"))]
+#[cfg(feature = "cortex-m-basepri")]
 pub use cortex_basepri::*;
 
 // Cortex-M target with source mask support
-#[cfg(any(feature = "cortex-m-source-masking", feature = "rtic-uitestv6"))]
+#[cfg(feature = "cortex-m-source-masking")]
 mod cortex_source_mask;
 
-#[cfg(any(feature = "cortex-m-source-masking", feature = "rtic-uitestv6"))]
+#[cfg(feature = "cortex-m-source-masking")]
 pub use cortex_source_mask::*;
 
-// RISC-V target (any)
 #[cfg(feature = "riscv")]
-pub use riscv_common::*;
+pub mod riscv_common;
 
 #[cfg(feature = "riscv")]
-mod riscv_common;
+pub use riscv_common::*;
 
 #[cfg(feature = "riscv-esp32c3")]
 mod riscv_esp32c3;
 #[cfg(feature = "riscv-esp32c3")]
 pub use riscv_esp32c3::*;
 
-#[inline(always)]
-pub fn assert_send<T>()
-where
-    T: Send,
-{
-}
+#[cfg(feature = "riscv-slic")]
+mod slic;
+#[cfg(feature = "riscv-slic")]
+pub use slic::*;
 
 #[inline(always)]
-pub fn assert_sync<T>()
-where
-    T: Sync,
-{
-}
+pub fn assert_send<T: Send>() {}
+
+#[inline(always)]
+pub fn assert_sync<T: Sync>() {}